Restructure flood_modeller.units module. Also some other bug-fixes.
Re-organised flood_modeller.units module to avoid the '_units' submodule.
Added Reservoir unit to valid units list.
Fixed line numbers to start from one and not zero.
Set default values for conduit top/bottom slot flags.
Fixed storage of reservoir rev1 lateral node labels.
Fixed name of reservoir row count field.
Fixed keyword validation function.
Fixed Rules unit parsing.
