+++ chyme/flood_modeller/io.py
@@ -390,11 +390,13 @@ class CircularConduitUnitIO(FloodModellerSubUnitIO):
             FloatDataField("invert", 0, 10),
             FloatDataField("diameter", 10, 10),
             StringDataField("bottom_slot_flag", 20, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("bottom_slot_height", 30, 10),
             FloatDataField("bottom_slot_depth", 40, 10),
             StringDataField("top_slot_flag", 50, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("top_slot_height", 60, 10),
             FloatDataField("top_slot_depth", 70, 10)]),
         DataRow([
@@ -420,11 +422,13 @@ class RectangularConduitUnitIO(FloodModellerSubUnitIO):
             FloatDataField("width", 10, 10),
             FloatDataField("height", 20, 10),
             StringDataField("bottom_slot_flag", 30, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("bottom_slot_height", 40, 10),
             FloatDataField("bottom_slot_depth", 50, 10),
             StringDataField("top_slot_flag", 60, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("top_slot_height", 70, 10),
             FloatDataField("top_slot_depth", 80, 10)]),
         DataRow([
@@ -451,11 +455,13 @@ class FullArchConduitUnitIO(FloodModellerSubUnitIO):
             FloatDataField("width", 10, 10),
             FloatDataField("height", 20, 10),
             StringDataField("bottom_slot_flag", 30, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("bottom_slot_height", 40, 10),
             FloatDataField("bottom_slot_depth", 50, 10),
             StringDataField("top_slot_flag", 60, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("top_slot_height", 70, 10),
             FloatDataField("top_slot_depth", 80, 10)]),
         DataRow([
@@ -482,11 +488,13 @@ class SprungArchConduitUnitIO(FloodModellerSubUnitIO):
             FloatDataField("springing_height", 20, 10),
             FloatDataField("arch_height", 30, 10),
             StringDataField("bottom_slot_flag", 40, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("bottom_slot_height", 50, 10),
             FloatDataField("bottom_slot_depth", 60, 10),
             StringDataField("top_slot_flag", 70, 10,
-                            valid_values = ['ON', 'OFF', 'GLOBAL']),
+                            valid_values = ['ON', 'OFF', 'GLOBAL'],
+                            blank_value = 'GLOBAL'),
             FloatDataField("top_slot_height", 80, 10),
             FloatDataField("top_slot_depth", 90, 10)]),
         DataRow([
@@ -547,8 +555,9 @@ class ReservoirUnitIO(FloodModellerUnitIO):
     components = [
         NodeLabelRow(count=0),
         NodeLabelRow(count=4,
+                     list_attribute_name="lateral_node_labels",
                      condition=lambda x: x.revision > 0),
-        DataRow([IntegerDataField("sa_row_count", 0, 10, apply_required=True)]),
+        DataRow([IntegerDataField("za_row_count", 0, 10, apply_required=True)]),
         DataTable("za", "za_row_count",
                   DataRow([
                       FloatDataField("z", 0, 10),
