+++ chyme/flood_modeller/units/structure.py
@@ -13,7 +13,7 @@
 
 """
 
-from ..units import FloodModellerUnit
+from .core import FloodModellerUnit
 
 class StructureUnit(FloodModellerUnit):
     def __init__(self, *args, io, **kwargs):
